refactor: organize evaluation as a package
