+++ src/evolve/evaluation/__init__.py
@@ -1,0 +1,17 @@
+from .results import (
+    CANONICAL_OUTCOMES,
+    EvaluationRecord,
+    Outcome,
+    TrialResult,
+    classify_evaluation,
+    evaluation_status,
+)
+
+__all__ = [
+    "CANONICAL_OUTCOMES",
+    "EvaluationRecord",
+    "Outcome",
+    "TrialResult",
+    "classify_evaluation",
+    "evaluation_status",
+]
