+++ src/evolve/evaluation/execution.py
@@ -8,13 +8,17 @@ import time
 from pathlib import Path
 from typing import Any
 
-from .config import evaluator_boolean, evaluator_sampling, experiment_id, load_config
-from .evaluation import EvaluationInterrupted, EvaluationRecord, Outcome, classify_evaluation
-from .git import evaluator_tree, git, git_stdout
-from .host_runtime import clean_python_env
-from .runtime import OwnedResult, attempt_dir, next_attempt, owned_attempt_id, run_owned
-from .task_sets import effective_task_set_identity
-from .task_vectors import trial_results, validate_task_vector
+from ..config import evaluator_boolean, evaluator_sampling, experiment_id, load_config
+from ..git import evaluator_tree, git, git_stdout
+from ..host_runtime import clean_python_env
+from ..runtime import OwnedResult, attempt_dir, next_attempt, owned_attempt_id, run_owned
+from ..task_sets import effective_task_set_identity
+from ..task_vectors import trial_results, validate_task_vector
+from .results import EvaluationRecord, Outcome, classify_evaluation
+
+
+class EvaluationInterrupted(BaseException):
+    """Carries a cancelled attempt to the driver for append-before-reraise."""
 
 
 def evaluate(
